chore(pyrefly): finalize global config, suppress unnecessary-type-conversion
- Suppress unnecessary-type-conversion (redundant bool/int/float/str calls
  that are idiomatic in PyTorch code, e.g. int(tensor.shape[-1]) for
  kernel launches) alongside existing third-party/dynamic-runtime suppressions
- Remove genuinely-redundant type-conversion calls from 9 files:
  bool(is_taichi_available()), bool(use_taichi after isinstance guard),
  int(f_dim // 2), float(value), int(reward_every_n_steps), str(decoded[0])
- Keep int(k.shape[-1]) in reinforced_kernels.py: needed for Taichi
  kernel dispatch (verified via test failure when removed)

Verified: pyrefly check 0 diagnostics, ruff clean, 530 tests pass.

+++ src/qwendopamine/models/gdn2_gpt/params.py
@@ -36,7 +36,7 @@ def compute_model_params(cfg: GDN2GPTConfig) -> dict[str, int]:
 
     k_dim = cfg.n_head * cfg.head_size
     hv = int(cfg.head_size * cfg.expand_v)
-    v_dim = int(cfg.n_head * hv)
+    v_dim = cfg.n_head * hv
     gdn2_projs = (cfg.n_embd * k_dim * 2) + (cfg.n_embd * v_dim)
     conv_params = (
         (k_dim * cfg.conv_size * 2) + (v_dim * cfg.conv_size)
